Move sphinx-related code into a dedicated module
